code/backend/services/client/requirements-test.txt
pytest==8.3.5
pytest-asyncio==0.26.0
Pillow
bcrypt
PyJWT
aiosqlite
SQLAlchemy
